实现余额计算

+++ src/main/java/cn/edu/usst/cs/campusAid/model/User.java
@@ -43,5 +43,5 @@ public class User {
      *     </ul>
      * </ul>
      */
-    float balance;
+    double balance;
 }
